feat: moved connection methods to core and deprecated existing ones
