+++ repository/repository.go
@@ -3,8 +3,9 @@ package e_repository
 import (
 	"context"
 	"errors"
-	"github.com/elcengine/elemental/connection"
 	"log"
+
+	elemental "github.com/elcengine/elemental/core"
 
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
@@ -19,7 +20,7 @@ func NewRepository[T any](collection string) Repository[T] {
 }
 
 func (r Repository[T]) Create(payload T) primitive.ObjectID {
-	result, err := e_connection.UseDefault().Collection(r.collection).InsertOne(context.TODO(), payload)
+	result, err := elemental.UseDefaultDatabase().Collection(r.collection).InsertOne(context.TODO(), payload)
 	if err != nil {
 		panic(err)
 	}
@@ -28,7 +29,7 @@ func (r Repository[T]) Create(payload T) primitive.ObjectID {
 
 func (r Repository[T]) FindOne(query primitive.M) *T {
 	model := new(T)
-	doc := e_connection.UseDefault().Collection(r.collection).FindOne(context.Background(), query)
+	doc := elemental.UseDefaultDatabase().Collection(r.collection).FindOne(context.Background(), query)
 	if doc.Err() != nil {
 		if errors.Is(doc.Err(), mongo.ErrNoDocuments) {
 			log.Fatalf("%v %s", r, doc.Err().Error())
@@ -46,7 +47,7 @@ func (r Repository[T]) FindByID(id primitive.ObjectID) *T {
 
 func (r Repository[T]) FindAll() []T {
 	var users []T
-	cursor, err := e_connection.UseDefault().Collection(r.collection).Find(context.Background(), primitive.M{})
+	cursor, err := elemental.UseDefaultDatabase().Collection(r.collection).Find(context.Background(), primitive.M{})
 	if err != nil {
 		panic(err)
 	}
@@ -55,14 +56,14 @@ func (r Repository[T]) FindAll() []T {
 }
 
 func (r Repository[T]) Update(id primitive.ObjectID, payload T) {
-	_, err := e_connection.UseDefault().Collection(r.collection).UpdateOne(context.Background(), primitive.M{"_id": id}, primitive.M{"$set": payload})
+	_, err := elemental.UseDefaultDatabase().Collection(r.collection).UpdateOne(context.Background(), primitive.M{"_id": id}, primitive.M{"$set": payload})
 	if err != nil {
 		panic(err)
 	}
 }
 
 func (r Repository[T]) Delete(id primitive.ObjectID) {
-	_, err := e_connection.UseDefault().Collection(r.collection).DeleteOne(context.Background(), primitive.M{"_id": id})
+	_, err := elemental.UseDefaultDatabase().Collection(r.collection).DeleteOne(context.Background(), primitive.M{"_id": id})
 	if err != nil {
 		panic(err)
 	}
