+++ tests/connection_test.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 	"testing"
 
-	e_connection "github.com/elcengine/elemental/connection"
 	e_constants "github.com/elcengine/elemental/constants"
+	elemental "github.com/elcengine/elemental/core"
 	e_mocks "github.com/elcengine/elemental/tests/mocks"
 	. "github.com/smartystreets/goconvey/convey"
 	"go.mongodb.org/mongo-driver/mongo/options"
@@ -15,29 +15,32 @@ import (
 func TestConnection(t *testing.T) {
 	Convey("Connect to a local database", t, func() {
 		Convey("Simplest form of connect with just a URI", func() {
-			client := e_connection.Connect(e_connection.ConnectionOptions{
-				URI: strings.Replace(e_mocks.DEFAULT_DATASOURCE, e_mocks.DEFAULT_DB_NAME, t.Name(), 1),
-			})
+			client := elemental.Connect(strings.Replace(e_mocks.DEFAULT_DATASOURCE, e_mocks.DEFAULT_DB_NAME, t.Name(), 1))
 			So(client, ShouldNotBeNil)
 			Convey("Should use the default database", func() {
-				So(e_connection.UseDefault().Name(), ShouldEqual, t.Name())
+				So(elemental.UseDefaultDatabase().Name(), ShouldEqual, t.Name())
 			})
 			Convey("Should use the specified database", func() {
 				DATABASE := fmt.Sprintf("%s_%s", t.Name(), "secondary")
-				So(e_connection.Use(DATABASE).Name(), ShouldEqual, DATABASE)
+				So(elemental.UseDatabase(DATABASE).Name(), ShouldEqual, DATABASE)
 			})
 		})
 		Convey("Connect with a URI specified within client options", func() {
 			opts := options.Client().ApplyURI(strings.Replace(e_mocks.DEFAULT_DATASOURCE, e_mocks.DEFAULT_DB_NAME, t.Name(), 1))
-			client := e_connection.Connect(e_connection.ConnectionOptions{
+			client := elemental.Connect(elemental.ConnectionOptions{
 				ClientOptions: opts,
 			})
 			So(client, ShouldNotBeNil)
 		})
 		Convey("Connect with no URI", func() {
 			So(func() {
-				e_connection.Connect(e_connection.ConnectionOptions{})
+				elemental.Connect(elemental.ConnectionOptions{})
 			}, ShouldPanicWith, e_constants.ErrURIRequired)
+		})
+		Convey("Connect with invalid argument", func() {
+			So(func() {
+				elemental.Connect(123)
+			}, ShouldPanicWith, e_constants.ErrInvalidConnectionArgument)
 		})
 	})
 }
