+++ tests/setup/setup.go
@@ -3,13 +3,13 @@ package e_test_setup
 import (
 	"strings"
 
-	e_connection "github.com/elcengine/elemental/connection"
+	elemental "github.com/elcengine/elemental/core"
 	e_test_base "github.com/elcengine/elemental/tests/base"
 	e_mocks "github.com/elcengine/elemental/tests/mocks"
 )
 
 func Connection(databaseName string) {
-	e_connection.ConnectURI(strings.Replace(e_mocks.DEFAULT_DATASOURCE, e_mocks.DEFAULT_DB_NAME, databaseName, 1))
+	elemental.Connect(strings.Replace(e_mocks.DEFAULT_DATASOURCE, e_mocks.DEFAULT_DB_NAME, databaseName, 1))
 }
 
 func Seed(databaseName string) {
@@ -22,6 +22,6 @@ func SeededConnection(databaseName string) {
 }
 
 func Teardown() {
-	e_connection.DropAll()
-	e_connection.Disconnect()
+	elemental.DropAllDatabases()
+	elemental.Disconnect()
 }
